requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
click==7.1.1
pathspec==0.7.0
Pillow==7.0.0
regex==2020.2.20
toml==0.10.0
typed-ast==1.4.1
